requirements.txt
mcp
fastapi>=0.110.0
uvicorn>=0.27.0
requests>=2.31.0
asyncpg>=0.29.0
typing-extensions>=4.0.0
python-dotenv>=1.0.0

